+++ src/test/java/net/snowflake/client/core/arrow/SmallIntToFixedConverterTest.java
@@ -35,7 +35,7 @@ public class SmallIntToFixedConverterTest extends BaseConverterTest
   /**
    * allocator for arrow
    */
-  private BufferAllocator allocator = new RootAllocator(Integer.MAX_VALUE);
+  private BufferAllocator allocator = new RootAllocator(Long.MAX_VALUE);
 
   /**
    * Random seed
