+++ src/test/java/net/snowflake/client/core/arrow/ThreeFieldStructToTimestampTZConverterTest.java
@@ -61,7 +61,7 @@ public class ThreeFieldStructToTimestampTZConverterTest extends BaseConverterTes
   /**
    * allocator for arrow
    */
-  private BufferAllocator allocator = new RootAllocator(Integer.MAX_VALUE);
+  private BufferAllocator allocator = new RootAllocator(Long.MAX_VALUE);
 
   private Random random = new Random();
