+++ src/test/java/net/snowflake/client/core/arrow/VarCharConverterTest.java
@@ -31,7 +31,7 @@ public class VarCharConverterTest extends BaseConverterTest
   /**
    * allocator for arrow
    */
-  private BufferAllocator allocator = new RootAllocator(Integer.MAX_VALUE);
+  private BufferAllocator allocator = new RootAllocator(Long.MAX_VALUE);
 
   private Random random = new Random();
